backend/requirements.txt
pipecat-ai[livekit,openai] 
python-dotenv 
loguru
fastapi
uvicorn[standard]
httpx
